+++ components/utils/src/templates.rs
@@ -56,11 +56,13 @@ pub fn get_shortcodes(tera: &Tera) -> HashMap<String, ShortcodeDefinition> {
 
         if template.name.starts_with("__zola_builtins/shortcodes/") {
             let head_len = "__zola_builtins/shortcodes/".len();
-            shortcode_definitions.insert(
-                identifier[head_len..(identifier.len() - ext_len - 1)].to_string(),
-                ShortcodeDefinition::new(file_type, &template.name),
-            );
-            continue;
+            let name = &identifier[head_len..(identifier.len() - ext_len - 1)];
+            // We don't keep the built-ins one if the user provided one
+            if shortcode_definitions.contains_key(name) {
+                continue;
+            }
+            shortcode_definitions
+                .insert(name.to_string(), ShortcodeDefinition::new(file_type, &template.name));
         }
     }
 
@@ -147,7 +149,7 @@ pub fn check_template_fallbacks<'a>(
 
 #[cfg(test)]
 mod tests {
-    use crate::templates::check_template_fallbacks;
+    use crate::templates::{check_template_fallbacks, get_shortcodes};
 
     use super::rewrite_theme_paths;
     use libs::tera::Tera;
@@ -192,5 +194,14 @@ mod tests {
             check_template_fallbacks("theme-only.html", &tera, &Some("hyde".to_string())),
             Some("hyde/templates/theme-only.html")
         );
+    }
+
+    #[test]
+    fn can_overwrite_builtin_shortcodes() {
+        let mut tera = Tera::parse("test-templates/*.html").unwrap();
+        tera.add_raw_template("__zola_builtins/shortcodes/youtube.html", "Builtin").unwrap();
+        tera.add_raw_template("shortcodes/youtube.html", "Hello").unwrap();
+        let definitions = get_shortcodes(&tera);
+        assert_eq!(definitions["youtube"].tera_name, "shortcodes/youtube.html");
     }
 }